+++ scripts/macros/2024/classFeatures/bard/dance/tandemFootwork.js
@@ -69,6 +69,11 @@ export let tandemFootwork = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ],
     config: [