+++ scripts/macros/2024/classFeatures/bard/fontOfInspiration.js
@@ -32,6 +32,11 @@ export let fontOfInspiration = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ]
 };