+++ scripts/macros/2024/classFeatures/bard/glamour/mantleOfInspiration.js
@@ -72,6 +72,11 @@ export let mantleOfInspiration = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ],
     config: [