+++ scripts/macros/2024/classFeatures/cleric/life/preserveLife.js
@@ -64,6 +64,11 @@ export let preserveLife = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ],
     config: [