+++ scripts/macros/2024/classFeatures/cleric/light/radianceOfTheDawn.js
@@ -52,6 +52,11 @@ export let radianceOfTheDawn = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ]
 };