+++ scripts/macros/2024/classFeatures/cleric/trickery/invokeDuplicity.js
@@ -183,6 +183,11 @@ export let invokeDuplicity = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ],
     config: [