+++ scripts/macros/2024/classFeatures/cleric/war/guidedStrike.js
@@ -66,6 +66,11 @@ export let guidedStrike = {
             pass: 'itemMedkit',
             macro: added,
             priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
         }
     ],
     config: [